refactor(research): extract semantic flow quality tests
